requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-docx>=1.1.0
